Refactored and added unit testing files
