+++ scripts/tests/test_data_transformations.py
@@ -1,0 +1,36 @@
+# Native
+import os
+import sys
+import unittest
+
+# Custom
+sys.path.append("..") # TODO Fix path.
+from utils.data_transformations import (convert_scientific_notation, merge_split_rows)
+
+# TODO Write unit tests.
+class TestDataTransformations(unittest.TestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        pass
+
+
+    @classmethod
+    def tearDownClass(cls):
+        pass
+
+
+    def setUp(self):
+        pass
+
+
+    def tearDown(self):
+        pass
+
+
+    def test_convert_scientific_notation(self):
+        pass
+
+
+    def test_merge_split_rows(self):
+        pass
