+++ scripts/tests/test_mdb_to_postgres.py
@@ -1,0 +1,52 @@
+# Native
+import os
+import sys
+import unittest
+
+# Custom
+sys.path.append("..") # TODO Fix path.
+from scripts.mdb_to_postgres import (get_tables
+                                    , get_table_columns
+                                    , generate_rows
+                                    , initialize_csv
+                                    , append_to_csv)
+
+# TODO Write unit tests.
+class TestMDBtoPostgres(unittest.TestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        pass
+
+
+    @classmethod
+    def tearDownClass(cls):
+        pass
+
+
+    def setUp(self):
+        pass
+
+
+    def tearDown(self):
+        pass
+
+
+    def test_get_tables(self):
+        pass
+
+
+    def test_get_table_columns(self):
+        pass
+
+
+    def test_generate_rows(self):
+        pass
+
+
+    def test_initialize_csv(self):
+        pass
+
+
+    def test_append_to__csv(self):
+        pass
