+++ scripts/tests/test_transfer_to_s3.py
@@ -1,0 +1,36 @@
+# Native
+import os
+import sys
+import unittest
+
+# Custom
+sys.path.append("..") # TODO Fix path.
+from scripts.transfer_to_s3 import (get_list_of_files, tempfile_to_s3)
+
+# TODO Write unit tests.
+class TestTransferToS3(unittest.TestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        pass
+
+
+    @classmethod
+    def tearDownClass(cls):
+        pass
+
+
+    def setUp(self):
+        pass
+
+
+    def tearDown(self):
+        pass
+
+
+    def test_get_list_of_files(self):
+        pass
+
+
+    def test_tempfile_to_s3(self):
+        pass
